Move to data.__init__
